Create HeartDisease
In this project, we use the Cleveland Heart Disease Dataset which consists of 14 features that are quantifiable. In the 14 features, the final feature is whether heart disease is occurring or not. So out of the 13 features, we take six features at a time and use combinatorics to generate 13C6 combinations which are a total of 1716 combinations. Four Classification algorithms which are Support Vector Machine, Random Forest, Decision Tree, K-Nearest Neighbours are applied on all these 1716 combinations to generate accuracy scores.

The arithmetic mean of all these accuracy scores are computed and each combination is assigned with that weighted mean score. We chose the top 100 combinations in terms of accuracy out of all the 1716 combinations. The feature which occurs the most in these top 100 combinations can be considered as the feature which is most helpful in predicting heart diease or the feature which is most positively correlated with the occurrance of heart disease.